Move mini USB connector and add solder jumpers
Unfortunately the mini USB connector was previously unuseable as it was
too far from the edge of the board (issue #28).  Moreover, when used as
a standard USB A the housing of a metal female USB A connector would
short against the pads of the unoccupied mini-USB connector if moved to
a useable position. For this reason solder jumpers are needed. These
jumpers will need to be shorted before the mini-USB port can be used.